+++ packages/fluent_ui_devtools_extension/lib/main.dart
@@ -12,9 +12,7 @@ class MainApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return const MaterialApp(
-      home: Scaffold(
-        body: HomePage()
-      ),
+      home: Scaffold(body: HomePage()),
     );
   }
 }