+++ packages/fluent_ui_devtools_extension/lib/src/pages/home_page.dart
@@ -7,7 +7,7 @@ class HomePage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return  DevToolsExtension(
+    return DevToolsExtension(
       child: IconsPage(),
     );
   }